Added GV_BOOT_UP
GV_BOOT_UP=true means Loop will know when it is running for the FIRST time.  It's also tested in the button pressed function due to it thinks the button is pressed on boot up for some reason.  button pressed will only do it's thing if GV_BOOT_UP is false.